NameRef and related improvements
Node class now handles parenting stuff, and you don't have to specify a
parent on construction. When your parent adds you, it sets itself as
your parent. This was required to make name-references work, which seem
to be more elegant now than the first go-round.

Also added full-service recursive import (imports in __init__.py's in
packages), and str methods on the primitives.